+++ apps/web/src/app/auth/page.tsx
@@ -1,63 +1,15 @@
 // apps/web/src/app/auth/page.tsx
-// 로그인 및 회원가입 페이지
-'use client';
+import { Suspense } from 'react';
+import LoginContents from './LoginContents';
 
-import { useEffect } from 'react';
-import {
-  useSearchParams,
-  useRouter,
-} from 'next/navigation';
-import KakaoLoginButton from '@/app/components/KakaoLoginButton';
-import { useAuthStore } from '@/stores/authStore';
-import { Route } from 'next';
-import LogoImage from '@/components/images/LogoImage';
-
-const ERROR_MESSAGES: Record<string, string> = {
-  kakao_login_failed:
-    '카카오 로그인에 실패했습니다. 다시 시도해주세요.',
-  session_expired:
-    '세션이 만료되었습니다. 다시 로그인해주세요.',
-  access_denied: '로그인이 취소되었습니다.',
-  no_authorization_code:
-    '인가 코드를 받지 못했습니다. 다시 시도해주세요.',
-  missing_code_verifier:
-    '보안 검증에 실패했습니다. 다시 시도해주세요.',
-  pkce_generation_failed:
-    'PKCE 생성에 실패했습니다. 브라우저를 새로고침 후 시도해주세요.',
-  invalid_request:
-    '잘못된 요청입니다. 다시 시도해주세요.',
-  kakao_auth_error:
-    '카카오 인증 서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요.',
-};
+export const metadata = {
+  title: '소소 로그인/회원가입',
+}; // 예시
 
 export default function AuthPage() {
-  const router = useRouter();
-  const params = useSearchParams();
-  const { isAuthenticated } = useAuthStore();
-
-  const errorCode = params.get('error');
-  const returnUrl = params.get('returnUrl');
-  const errorMessage = errorCode
-    ? ERROR_MESSAGES[errorCode]
-    : null;
-
-  useEffect(() => {
-    if (isAuthenticated) {
-      const to = returnUrl
-        ? decodeURIComponent(returnUrl)
-        : '/signup';
-      router.push(to as Route<string>);
-    }
-  }, [isAuthenticated, returnUrl, router]);
-
   return (
-    <div>
-      <div className="flex flex-col items-center justify-center ">
-        소소에 오신 것을 환영합니다!
-        <LogoImage />
-      </div>
-      {errorMessage && <p>{errorMessage}</p>}
-      <KakaoLoginButton />
-    </div>
+    <Suspense fallback={null}>
+      <LoginContents />
+    </Suspense>
   );
 }
